feat: Add mapping for latest winners in MonoNFT contract
This commit adds a mapping `_latestWinners` to keep track of the latest winners in the MonoNFT contract. The mapping is used to store the `Winner` struct, which contains the address of the winner, the price they paid, and the expiration date of their ownership rights. The `confirmWinner` function has been modified to update the `_latestWinners` mapping with the winner's details. Additionally, a new function `confirmWinner` has been added to allow specifying the expiration date externally for flexibility.